requirements.txt
uvicorn==0.38.0
google-genai==1.50.1
fastapi==0.121.2
python-dotenv==1.2.1
langchain==0.3.0
langchain-community==0.3.0
faiss-cpu==1.10.0
sentence-transformers>=2.7.0
huggingface-hub>=0.20.0